+++ directory/directory.go
@@ -1,4 +1,4 @@
-package main
+package directory
 
 import (
 	"fmt"
