+++ docker/docker_client.go
@@ -1,4 +1,4 @@
-package main
+package docker
 
 import (
 	"crypto/tls"
