+++ docker/docker_image_dest.go
@@ -1,4 +1,4 @@
-package main
+package docker
 
 import (
 	"bytes"
