+++ docker/docker_image_src.go
@@ -1,4 +1,4 @@
-package main
+package docker
 
 import (
 	"fmt"
