+++ openshift/openshift.go
@@ -1,4 +1,4 @@
-package main
+package openshift
 
 import (
 	"bytes"
@@ -13,6 +13,7 @@ import (
 	"strings"
 
 	"github.com/Sirupsen/logrus"
+	"github.com/projectatomic/skopeo/docker"
 	"github.com/projectatomic/skopeo/dockerutils"
 	"github.com/projectatomic/skopeo/types"
 	"github.com/projectatomic/skopeo/version"
@@ -232,7 +233,7 @@ func (s *openshiftImageSource) ensureImageIsResolved() error {
 		return err
 	}
 	logrus.Debugf("Resolved reference %#v", dockerRef)
-	d, err := NewDockerImageSource(dockerRef, s.certPath, s.tlsVerify)
+	d, err := docker.NewDockerImageSource(dockerRef, s.certPath, s.tlsVerify)
 	if err != nil {
 		return err
 	}
@@ -257,7 +258,7 @@ func NewOpenshiftImageDestination(imageName, certPath string, tlsVerify bool) (t
 	// i.e. a single signed image cannot be available under multiple tags.  But with types.ImageDestination, we don't know
 	// the manifest digest at this point.
 	dockerRef := fmt.Sprintf("%s/%s/%s:%s", client.dockerRegistryHostPart(), client.namespace, client.stream, client.tag)
-	docker, err := NewDockerImageDestination(dockerRef, certPath, tlsVerify)
+	docker, err := docker.NewDockerImageDestination(dockerRef, certPath, tlsVerify)
 	if err != nil {
 		return nil, err
 	}
